Bit-buffer + type migration to floats (#8)
Bit buffer and double to float migration

+++ src/main/java/records/DataPoint.java
@@ -1,4 +1,3 @@
 package records;
 
-public record DataPoint(long timestamp, double value) {
-}
+public record DataPoint(long timestamp, float value) {}
